chore(projects): refactor web project views for mobile
Initial refactor to bring in web project views to mobile and reuse that functionality.

We still need to fix the mobile task search functionality and modularize some of the code.
